+++ lib/attributor.rb
@@ -6,44 +6,6 @@ require 'hashie'
 require 'digest/sha1'
 
 module Attributor
-  # Checks if a given constraint is true for the current Ruby version
-  #
-  # @param [String] op A binary comparison operator
-  # @param [String] rhs A version to compare to the current Ruby version
-  #
-  # @example Basic usage
-  #   if Attributor.check_ruby_version('<', '1.9')
-  #     # Do something supported in Ruby 1.8 or lower
-  #   end
-  #
-  # @return [Boolean] true if "lhs op rhs" is true, false otherwise
-  #
-  def self.check_ruby_version(op, rhs)
-    self.compare_versions(RUBY_VERSION, op, rhs)
-  end
-
-  # Performs a comparison on two versions
-  #
-  # @param [String] lhs A String version that is the left-hand side of binary op
-  # @param [String] op A binary comparison operator
-  # @param [String] rhs A String version that is the right-hand side of binary op
-  #
-  # @example Basic usage
-  #   if Attributor.compare_versions(RUBY_VERSION, '<', '1.9')
-  #     # Do something supported in Ruby 1.8 or lower
-  #   end
-  #
-  # @return [Boolean] true if "lhs op rhs" is true, false otherwise
-  #
-  def self.compare_versions(lhs, op, rhs)
-    Gem::Version.new(lhs.dup).send(op.to_sym, Gem::Version.new(rhs.dup))
-  end
-
-  if Attributor.check_ruby_version('<', '1.9')
-    # Method require_relative was added in Ruby 1.9, so use the gem
-    # if we have a Ruby version less than 1.9
-    require 'require_relative'
-  end
 
   require_relative 'attributor/exceptions'
   require_relative 'attributor/attribute'
@@ -51,7 +13,7 @@ module Attributor
   require_relative 'attributor/dsl_compiler'
   require_relative 'attributor/attribute_resolver'
 
-  require_relative 'attributor/extensions/core'
+  
   require_relative 'attributor/extensions/randexp'
 
   require_relative 'attributor/types/container'
@@ -65,6 +27,8 @@ module Attributor
   require_relative 'attributor/types/float'
   require_relative 'attributor/types/collection'
   require_relative 'attributor/types/hash'
+  require_relative 'attributor/types/csv'
+  require_relative 'attributor/types/ids'
 
   # List of all basic types (i.e. not collections, structs or models)